requirements.txt
python-telegram-bot==20.3
requests==2.31.0
flask==2.2.5
gTTS==2.3.2
